requirements.txt
ultralytics==8.2.0
torch
gradio
pillow
reportlab
